requirements.txt
contractions==0.1.73
datasets==3.1.0
discord.py==2.4.0
importlib-metadata==8.0.0
ipykernel==6.29.5
jaraco.collections==5.1.0
nlpaug==1.1.11
nltk==3.9.1
nvidia-cuda-nvcc-cu12==12.5.82
onnx==1.17.0
onnxruntime==1.20.0
pysocks==1.7.1
python-dotenv==1.0.1
scikit-learn==1.5.2
seaborn==0.13.2
tensorflow==2.18.0
tomli==2.0.1
torchvision==0.20.1
transformers==4.46.2
